feat: add empty state to the packages page
