+++ static/js/store/components/PackageFilter/index.ts
@@ -1,0 +1,1 @@
+export { PackageFilter } from "./PackageFilter";
