+++ static/js/store/components/SearchInput/index.ts
@@ -1,0 +1,1 @@
+export { SearchInput } from "./SearchInput";
