+++ static/js/store/types/components/Store.d.ts
@@ -1,0 +1,11 @@
+import type { Category } from "./Category";
+import Package from "./Package";
+
+type Store = {
+  total_items: number;
+  total_pages: number;
+  packages: Package[];
+  categories: Category[];
+};
+
+export default Store;
