+++ static/js/store/types/index.d.ts
@@ -1,4 +1,5 @@
 import Category from "./components/Category";
 import Package from "./components/Package";
+import Store from "./components/Store";
 
-export { Category, Package };
+export { Category, Package, Store };
